move track gestures from TrackRenderingContainer to BlockBasedTrack
